+++ aom/cmd/ls.go
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/philippgille/apiomat-go/aomc"
-	"github.com/philippgille/apiomat-go/aoms"
+	"github.com/philippgille/apiomat-go/aomx"
 )
 
 // lsCmd represents the ls command
@@ -20,7 +20,7 @@ ls is a command for listing classes of a given module.
 For example, to list all classes of module "MyModule":
 aom class ls --module "MyModule" --baseUrl "https://apiomat.yourcompany.com/yambas/rest" --username "john" --password "secret"`,
 	Run: func(cmd *cobra.Command, args []string) {
-		client := aomc.NewDefaultClient(baseUrl, username, password, aoms.System(system))
+		client := aomc.NewDefaultClient(baseUrl, username, password, aomx.System(system))
 
 		classes, err := client.GetClasses(module)
 		if err != nil {